Fix issues with compiling with clang with c++11 support. Mostly issues with "user defined literals". Also some auto-casting no longer works.

+++ servers/audio/audio_mixer_sw.cpp
@@ -287,8 +287,8 @@ void AudioMixerSW::mix_channel(Channel& c) {
 		//stereo pan
 		float pan = c.pan * 0.5 + 0.5;
 		float panv[2]={
-			(1.0 - pan)*(1<<MIX_VOL_FRAC_BITS),
-			(pan)*(1<<MIX_VOL_FRAC_BITS)
+                        float((1.0 - pan)*(1<<MIX_VOL_FRAC_BITS)),
+			float((pan)*(1<<MIX_VOL_FRAC_BITS))
 		};
 
 		for(int i=0;i<2;i++) {
@@ -304,10 +304,10 @@ void AudioMixerSW::mix_channel(Channel& c) {
 		float pany = c.depth * 0.5 + 0.5;
 		// with this model every speaker plays at 0.25 energy at the center.. i'm not sure if it's correct but it seems to be balanced
 		float panv[4]={
-			(1.0-pany)*(1.0-panx)*(1<<MIX_VOL_FRAC_BITS),
-			(1.0-pany)*(    panx)*(1<<MIX_VOL_FRAC_BITS),
-			(    pany)*(1.0-panx)*(1<<MIX_VOL_FRAC_BITS),
-			(    pany)*(    panx)*(1<<MIX_VOL_FRAC_BITS)
+                        float((1.0-pany)*(1.0-panx)*(1<<MIX_VOL_FRAC_BITS)),
+			float((1.0-pany)*(    panx)*(1<<MIX_VOL_FRAC_BITS)),
+			float((    pany)*(1.0-panx)*(1<<MIX_VOL_FRAC_BITS)),
+			float((    pany)*(    panx)*(1<<MIX_VOL_FRAC_BITS))
 		};
 
 		for(int i=0;i<4;i++) {
